RAG/项目二_知识图谱RAG/requirements.txt
# 基础依赖
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pymupdf==1.23.8
markdown==3.5.1
python-dotenv==1.0.0
openai==1.3.0
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3

# 知识图谱特定依赖
neo4j==5.14.0
langchain-experimental==0.0.49
spacy==3.6.1
networkx==3.1
